fix: delete unuse posting state and modify error handling
