+++ components/comment/Form.tsx
@@ -2,34 +2,42 @@
 
 import { useState } from 'react';
 import { Button } from '../controls';
+import { createComment } from '@/app/actions/comment';
 
 interface FormProps {
-    onSubmit: (content: string) => Promise<void>;
-    isPosting: boolean;
+    comicId: string;
 }
 
-export default function Form({ onSubmit, isPosting }: FormProps) {
+export default function Form({ comicId }: FormProps) {
     const [newComment, setNewComment] = useState('');
 
-    const handleSubmit = async (e: React.FormEvent) => {
-        e.preventDefault();
-        if (!newComment.trim() || isPosting) return;
+    const handleAction = async () => {
+        if (!newComment.trim()) return;
 
-        await onSubmit(newComment.trim());
-        setNewComment('');
+        try {
+            const result = await createComment(comicId, newComment);
+            if (result?.error) {
+                alert('コメントの投稿に失敗しました。もう一度お試しください。');
+                return;
+            }
+            setNewComment('');
+        } catch (error) {
+            console.error('Failed to create comment:', error);
+            alert('コメントの投稿に失敗しました。もう一度お試しください。');
+            return;
+        }
     };
 
-    const isDisabled = !newComment.trim() || isPosting || newComment.length > 500;
+    const isDisabled = !newComment.trim() || newComment.length > 500;
 
     return (
-        <form onSubmit={handleSubmit} className="mb-6">
+        <form action={handleAction} className="mb-6">
             <textarea
                 value={newComment}
                 onChange={(e) => setNewComment(e.target.value)}
                 placeholder="コメントを入力してください..."
                 className="w-full p-3 border border-gray-300 resize-none focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-transparent"
                 rows={3}
-                disabled={isPosting}
             />
             <div className="flex justify-between items-center mt-2">
                 <span className="text-sm text-gray-500">
@@ -39,7 +47,7 @@ export default function Form({ onSubmit, isPosting }: FormProps) {
                     type="submit"
                     disabled={isDisabled}
                 >
-                    {isPosting ? '投稿中...' : '投稿する'}
+                    投稿する
                 </Button>
             </div>
         </form>
