+++ components/comment/Section.tsx
@@ -1,9 +1,9 @@
 'use client';
 
 import { useState, useEffect } from 'react';
-import { getComments, postComment, type Comment } from '../../lib/commentApi';
-import CommentForm from './Form';
-import CommentList from './List';
+import { getComments, type Comment } from '@/lib/commentApi';
+import Form from './Form';
+import List from './List';
 
 interface SectionProps {
     comicId: string;
@@ -12,7 +12,6 @@ interface SectionProps {
 export default function Section({ comicId }: SectionProps) {
     const [comments, setComments] = useState<Comment[]>([]);
     const [loading, setLoading] = useState(true);
-    const [posting, setPosting] = useState(false);
 
     useEffect(() => {
         fetchComments();
@@ -30,24 +29,11 @@ export default function Section({ comicId }: SectionProps) {
         }
     };
 
-    const handleCommentSubmit = async (content: string) => {
-        try {
-            setPosting(true);
-            await postComment(comicId, content);
-            await fetchComments();
-        } catch (error) {
-            console.error('コメント投稿エラー:', error);
-            alert('コメント投稿中にエラーが発生しました');
-        } finally {
-            setPosting(false);
-        }
-    };
-
     return (
         <div className="bg-white">
             <div className="p-4">
-                <CommentForm onSubmit={handleCommentSubmit} isPosting={posting} />
-                <CommentList comments={comments} loading={loading} />
+                <Form comicId={comicId} />
+                <List comments={comments} loading={loading} />
             </div>
         </div>
     );
